Renames lib to services
